Did tons of work on the weapon switching system
I figured out a method that may not be the most efficient, but it seems to work pretty well. I have multiple classes that contain separate groups of weapon stats, such as FireControlData, ProjectileData, OpticsData and so on. Each weapon has a selection of these scripts, and there is a firing mode class that references the appropriate stats. This also allows certain firing modes to share stats, such as a selective-fire rifle feeding from the same magazine.