THE FIRING MODES WORK!
The firing modes work! Switch between them with the scroll wheel.

There are still issues to sort out, though. For example, when switching from a mode with optics to one without, the gun stops updating the optics and stays stuck zoomed in until the player switches to a mode with optics. Additionally, values are not programmed in for unique projectile types, meaning that it is more difficult to change unique stats for different projectiles.

Also, I added a firing noise and a greatly improved muzzle flash system.

Plus, I created a more efficient system to move the weapon's model around based on a variety of actions.