Created a fully functioning weapon wheel
The weapon wheel allows the player to switch between each firing mode for each weapon in their arsenal. The single selection menu makes it very easy for the player to switch to any mode from any other mode, with only one button press and a movement of the mouse/analog stick.