Further tested code for taking cover, and completed directional dodge function
The take cover function is still somewhat buggy. I still need to figure out a null state for it.

I have not tested the dodge function, and it may not work. I also need to make a function for dodging area of effect attacks, and figure out an appropriate way to trigger it in code.