Implemented a seemingly better way to manage gun stats
As part of my process to make the weapon system better, I created a new 'Gun' MonoBehaviour. This started off as a copy of the old 'RangedWeapon' script and will slowly be turned into a cleaner, better version.
My first task was to overhaul the stat system. My old system consisted of multiple arrays of classes for different types of stats, and the data was stored in an additional FiringMode class, with index integers used to find the appropriate stat classes in the arrays. If that didn’t make any sense, then consider that an indicator of how janky and awful it was. The reason for this was to have multiple firing modes with different stats, but which could also share stats for certain variables that may need to be referenced between firing modes, such as the magazine capacity for a selective-fire rifle (as in both modes the gun is still using the same ammunition from the same magazine).
My new system instead makes the stat classes into MonoBehaviours, and assigns them to otherwise empty gameobjects that are children of the weapon prefab. This way, I can drag references to them directly onto the firing mode class. This may be cleaner for organising variables, but I’m not entirely sure yet, as transferring the data over was annoying, and I feel that most of the janky code in the weapon script might be from other functions. Either way, I intend to work on it more.
Possible fix to make it better:
Leave fire controls, magazine and optics alone, but combine recoil, accuracy, projectile, ammunition and cosmetics into a single class. As the latter five classes will almost always be used together, this will reduce the amount of messy code.