Started work on a new and improved reload handler
As part of my current task to overhaul my weapon system into a less terrible one, I have started making a new reload system. I learned some more information about Coroutines, most notably that they can be referenced, and if so, they can be stopped remotely. This will allow me to use them in a much more flexible manner.
The new reload sequence consists of two main functions - a regular void named ReloadHandler(), and an IENumerator named ReloadSequence(). ReloadHandler() is used for controlling the reload function, by detecting if the player has pressed the reload function or if the magazine is empty, and executing ReloadSequence(). ReloadSequence() is much simpler than I anticipated it would be. My old reload system requires counting up timers, to check if the cycle has finished. But with this system, instead of having to count up a timer, it simply uses WaitForSeconds() to wait for the duration of the reload time. I thought I needed to have a while loop based on a timer, so it could check if it had been disabled, but thanks to StopCoroutine() I do not need to do this. I have not tested this new system yet, but I am hopeful that it will be far superior.