Update - 09/11/2020
Tested and implemented new reload handler
The new reload IEnumerator system works perfectly, while being a lot cleaner and simpler than the previous system. One stumbling block I ran into was detecting remotely if the reload function was running, which I initially achieved using a bool. I realised I could remove this and simply check if the IEnumerator was null or not, but I found out that the function would remain present, even after it had ended. So the game would think the gun was still reloading. I needed to figure out a way to have the reload sequence IEnumerator automatically null itself on completion. My solution was to take the function I used for remotely cancelling the reload, which both stops and nulls it, and simply run it at the end of the reload cycle. There is no harm done from not letting an IEnumerator finish, and by having it cancel itself instead of ending normally, it nulls itself without causing any side effects. All in all, I am extremely satisfied with my new system.

Fixed grenades
The grenades had strangely weak damage, and I found out that the cause was that not all of the stats were changed properly from the old system. I changed the grenade stats and rectified this.

I also attempted to fix the ADS code, to solve a jittering problem. This did not succeed. I also somehow broke the muzzle flash cosmetic effect for the player's guns, and have unfortunately not been able to figure out the cause.