Changelog 16/11/2020
Fixed the muzzle flash effect
I had the player's guns' muzzle flash effects disabled for ease of working with while in the scene, but they were not coded to re-enable when the game was loaded. Therefore they would not run in-game.

Resized the UI to a correct aspect ratio
I originally planned to have the game in pixel art, and resized the UI to match. Eventually I decided to switch this. The UI was still a weird tiny size, though, so I changed the HUD canvas to a larger sensible aspect ratio, and resized everything accordingly.