Changelog 17/11/2020
Fixed the ADS model lerping code.

As I have been overhauling the player's gun code, I have also obviously been reworking the ADS code. I tried setting up a better system for moving the gun to an appropriate position so the weapon's sightline lines up with the player's vision. This caused the gun to glitch out. I realised a problem was that I was using the wrong transform as reference for calculating appropriate values, which was causing the position and rotation to jitter about. I also consulted a forum to obtain the best advice for calculating the rotations, and the code people gave me worked beautifully.

Removed some miscellaneous debug messages so they wouldn't clutter up the console.