Changelog 18/11/2020
Started creating a better function for cancelling the ADS process.

I might change the ADS function to use an IEnumerator to enable the ADS when pressed down, and another to disable it when the button is released. Or one function with a variable to specify. Either way, this might be better to ensure it is cancelled properly when switching weapons or reloading.