Changelog 23/11/2020
Performed more work on the options menu. It's been a while since I've made an options menu, so since I've been having so much trouble I opted to find one from one of my previous project. Unfortunately, this went nowhere, as the one I made previously appeared to be based off the exact same tutorial I'm currently going off of.

Despite this, I still managed to complete a basic options menu. While I have not tested it yet, the code is very clean and I'm quite proud of it. Due to the sheer number of individual options I had to repeatedly code for, with no good way I could think of to automate it, I found it difficult organising all my code to make it easy to read. But I think I've managed for now.