Changelog 24/11/2020
Performed limited testing on the options menu. In case I didn't mention it last time, I completed a very basic but functional one. It seems to load and display the options correctly, but I don't know very well if it's actually applying settings. I know that the audio volume is not changing, for example.

Additionally, I found out that the player's firing mode selector on their weapons would still register even when the game was paused. I fixed this with a simple if statement in Update() of the weapon script, that ensures that Time.timeScale is not zero before running anything else.

I also experimented with a surprisingly simple system for menu hierarchies, involving a struct containing a MenuWindow reference, plus an array of the struct itself, to allow for an infinite amount of branching children. This initially felt very clever, but looked very messy in the Inspector. Additionally, I realised it would be simpler to achieve the same thing in the MenuWindow class itself, without an additional struct.

I also removed some old experimental menu scripts I decided to scrap.