Changelog 02/12/2020
Fixed a bug that prevented the player from returning to previous menu screens.
For some reason, menu window items would not be able to recognise their MenuHandler. I performed a quick fix by having the function for returning to a previous menu find the MenuHandler in their parent on the fly. This method is less performant, but it actually works without any fuss.