Changelog 08/12/2020
Began some work creating an improved system to play repeating or looping sound effects. One problem with repeating the same effect over and over is that an uncanny-valley-esque effect sets in. In real life, when performing the same noisy action multiple times, there will often be subtle differences in the way an action is performed that results in a slightly different sound. E.g. different footsteps due to stepping on different materials or bodyweight causing the person's foot to hit the ground at a different speed and angle. I intend on trying several methods of accomplishing this:

* Randomly selecting one from several similar sounds

* Having one sound and randomly altering the pitch by a small margin in real time

* Having a single loop of different sounds