Changelog 15/12/2020
After a short hiatus to focus on assignments, I'm back, it's the end of the year and I'm burnt out. To keep my momentum going, I'm attempting to fix one of the few things I can actually summon the energy and enthusiasm to work on right now.

The code that the enemy AI uses for aiming and attacking is somewhat poorly implemented. The enemy will slowly aim towards the player instead of immediately snapping to position, a feature I implemented to allow the player to dodge near-instant attacks like bullets. However, the current system has a fixed point in space that moves towards the target at a constant speed, and which the enemy aim snaps to. This poses problems as if the player is rapidly moving towards or away from the enemy, their aim will take a strangely long time to catch up, even if the angle is very small. And the speed doesn’t account for distance either. If I adjust an enemy’s aim speed so they are effective at long range, they aim far too quickly at short range. If I lower it so they can be dodged at short ranges, they become sluggish at long range.
I wish to change this so the enemy simply rotates to face the target at a certain angle per second, which will ensure a constant rotational speed regardless of distance. While the old system could be useful for some enemies (e.g. a gunner in a perch, who must be strafed at close range to keep out of harm’s way), I will probably be able to implement such a mechanic into the new system while also having the new advantages.

My enemy aiming code is also messily mixed up with the actual attack animations. To fix this, I opted to make dedicated functions within the AI class – LookTowards() to move the enemy’s aim around, and IsLookingAt() to determine if the enemy is within a certain angle threshold. These functions can be cleanly referenced by any other scripts that want to have the enemy look at things.