Update 18/03/2021
It's been a while since I worked on this!

Finally created a basic interaction system. When the player looks at a collider bearing the ‘Interactable’ script, a window will display showing the object’s name and a prompt. Pressing the ‘Interact’ button (currently bound to E) will invoke a UnityEvent doing whatever is specified in the Interactable, and if necessary, creating a short cooldown before the object can be interacted with again. This still needs work however, such as adding a visual button prompt. I tried to figure out how to get an input and find the button assigned to it, but this proved futile.
Started creating a new heads up display script, named ‘HUD’. My old script is getting too cluttered, and relies on constantly refreshing HUD elements in Update(), which is bad for performance. I intend my new system to listen for when the player’s status changes, using an EventObserver, and change things from there. So for example, it will only change the health meter when the player’s health actually changes.
Created constructors for the various event message classes used for my EventObserver system, making it easier to create new ones in other scripts.
Started work on a class called ‘KillPerimeter’, to replace my old script ‘KillBox’. This will be a MonoBehaviour that exists in the centre of a level, and at regular intervals, sweep for entities. It will immediately kill, repool or otherwise destroy anything that’s a certain distance away from it. This is meant to deal with any gameObjects that fall out of the level, and I theorise it will work much better than my old ‘KillBox’ script. Said script relied on colliders, and as a result is very unstable for objects moving at high speeds, which will often bypass the collider and continue falling forever.
Cleaned up my base AI script by deleting some unnecessary Line of Sight functions, and moving several general purpose functions to the ‘AIFunction’ class where they belong.
Started overhauling the attacking system for AI. I created two new scripts with placeholder names, ‘FightingAI’ and ‘EntityAttack’. The former inherits from ‘AI’ and will be a new variant of ‘Combatant’, and ‘EntityAttack’ will be a new version of ‘NPCAttack’.