Update from 26/04/2021 to 29/04/2021
Added ammo box model, and started work on making interactable ammo boxes that will fully refill all ammunition of a particular type.

Added door model, and started work on making an interactable door that can be opened and closed.

Worked on interaction system. I’m trying to figure out a way to have an interactable object change between different prompts. I’ve been debating multiple options to try and figure out the best one.
One thought I was suggested was to have a separate Interaction struct, which the Interactable monobehaviour has an array of. This proved unnecessarily complicated and visually difficult to process, as most objects would only have one prompt. My current plan is to have multiple separate Interactable functions on one object, and have a way to enable and disable them as necessary. This may pose other problems, as it will make it harder for the player’s interaction function to select the right one. That being said, I feel this may be easy to solve, and it is the method I will most likely go with.