Update 30/04/2021 - Attempted to fix door, added denial criteria to Interactable
I did more work trying to make my door controls work correctly, but it hasn't helped.

I added an extra feature to my Interactable script, a virtual bool named CanPlayerInteract(PlayerHandler ph). This bool will determine whether a player can interact with an object. This is a virtual bool so that I can make inherited classes for certain interactable objects, with unique criteria to determine whether the player can interact with it. For example, an ammo box can check the player's ammunition inventory, and if their ammunition is full it will not activate. Or if a door is locked, it will check if the player has the key before allowing itself to be opened.