Update - 03/05/2021 - 05/05/2021, Finished ammo boxes, door breakthrough, walled off test level, slightly improved interaction window, and an out-of-bounds kill function.
Made my ammo box work by creating a new script named AmmunitionRefillInteractable, deriving from Interactable. The script functions similarly to the base Interactable class, but contains a reference to an AmmoPickup monobehaviour on the same object. The interaction function will deposit ammunition into the player’s inventory, a function that cannot be specified on the fly within the default UnityEvent. Additionally, it alters the CanPlayerInteract bool, to check if the player’s ammunition is not full, before allowing them to interact. The ammo boxes work completely now (save for a tiny error in the actual animation, that causes components to clip into each other.
Made separate prefabs for ammo boxes that refill the player’s four main ammunition types, with unique textures to match. There was only a texture for a rifle ammunition box, but I added 3 new textures for pistol bullets, shotgun shells and impact grenades. I have not made a prefab or texture for shotgun slugs, but I may remove this ammunition type in the future.

I decided to take a new angle when making my door interface work properly. Instead of having two different interaction prompts with different tooltips, I realised I could make do with just one.
I created a new function named DoorInteractable, that derives from Interactable. This script combines an Interactable class with my old DoorHandler class, which is now unnecessary. This script contains a reference to an animator controller, and when activated will reference and toggle the bool, triggering the appropriate opening or closing animation. It also modifies the CanPlayerInteract bool, to account for if the door is locked or not.
This system worked! For a bit. I tried moving the animator controller to a different object in the hierarchy, and it didn’t work. I changed it back, and it still doesn’t work. The transitions, controls and interaction window all work as intended, but the animations themselves do not. I checked if my states in the animator controller did not have animations assigned to them, but that was not the case. I will need to figure this out.

Added a bool to the ItemPickup class, specifying if an item is automatically picked up on contact. If false, the player will need to directly interact with it. This function is not fully completed yet, but it works as a stopgap to allow my ammo boxes to work without being destroyed if the player bumps into them.

Altered the level layout by adding perimeter walls, and another small enclosed building with a door.

Changed the new HUD script’s name to NoUpdateHUD, to differentiate it from the heads up display script currently in use.

Created extra indicator sprites for the interaction window in the HUD, to show if an interactable object is accessible, denied or in the progress of performing a function upon interaction.
Changed some of the dimensions of text recttransforms in the heads up display’s interaction window, to add some whitespace for easier readability and aesthetic appeal.

Created a MonoBehaviour named KillRadius, that automatically destroys any entity that falls out of bounds. This will be more effective than having kill colliders, since with those there’s a possibility of the collision detection failing.