Update 07/05/2021 - 08/05/2021 - KillRadius test, finished door function
Attempted to implement the KillRadius function in a scene and test it with the player, but it didn't work. I have more important things to work on, though.

I figured out the problem with the door controller. Some of the names of the objects in the hierarchy for the door model were renamed, and this prevented the animation from working, presumably because the animation data couldn't find the correct transforms to move. Due to the terrible names of the individual objects, which I wanted to change, I had to re-export my door model. This also let me remove unnecessary elements such as lights and cameras, and ensure that certain elements were exported separately, something that I forgot to do last time. But now the door works. I still need to add audio to it and the ammo box though