Update 23/05/2021: Adding some sounds to door and ammo box, bugfixes
After attempting to add sound effects to the ammo box and door via animation events, which I couldn’t figure out how to do, I decided I had more important things to spend my time and effort on. I opted to simply run the sound effects via the UnityEvents in their Interactable scripts, directly calling the function whenever the player interacts with them.

I found the AudioSources for a lot of sound effects in the game, and changed them so they ran using the Sound Effects audio mixer. This will be necessary so the player can adjust the volume of sound effects accordingly, and separately from other sounds like music and dialogue.

I found that if the player starts with less ammunition for a gun than its magazine will carry, the HUD counter will bug out because the magazine has a higher capacity than how much ammunition there is. I added a simple check for this, that runs whenever the Gun script is enabled, to adjust the magazines accordingly.

I changed most of the ammo box prefabs to be prefab variants rather than separate prefabs – this will make it easier to make changes in the future.

I buffed the Gunman enemy to fire faster and attack more aggressively, so it poses more of a threat. That being said, their awkward aiming system makes it very hard for them to establish a lock on the player, so they still aren’t very dangerous. I need to alter their aiming to work on rotation rather than following a marker, since it’s too easy for the player to move away from the marker.