Update 25/05/2021: Fixed bug with weapon magazines
Fixed a bug where the weapons would have empty magazines when the scene loads, even if the player has ammunition. The ammo indicators on the weapon wheel do not seem to display the amounts properly, but this is minor and will need to be fixed later on as opposed to soon.