Update 01/06/2021 - 02/06/2021 - Random sound playing and a potential new weapon
Started creating a RandomSoundPlayer ScriptableObject. This object contains a list of random sounds, plus options for pitch variance, and a delay option. This will make it easier to play sounds, in any scenario where a sound will need to have several random variants. I should be able to drag one of these into a UnityEvent, and run its Play(AudioSource source) function using the appropriate audiosource for the object. Unfortunately, I have not been able to test it yet, and I also need to have functionality to change the audiosource’s pitch back to whatever it was before it’s changed using the random pitch variance.

Started creating a flamethrower weapon. This weapon was originally going to be added, but then cut due to the setting, but then I decided to add it as a way to test some mechanics, such as the random sound player (because I made some fireball noises for someone else). I ended up deciding not to, though, so I'm not really sure where I'm going with this.

I tried re-adding the Petrol ammo type to the game, but I found out that my ammunition inventory script does not re-organise itself to match. I have barely changed anything about this script, so it’s baffling that it does not work anymore. I will need to fix it at another time.