Update 24/06/2021 - Started work on a new AI attack system
Since my enemies still have many problems that I need to work out, and my code is old and messy enough that modifying that is going to be annoying, I have elected to create a brand new overhauled enemy attack system. I started this process a while ago but had other obligations to deal with so I had to hold it off.