Update 02/05/2021 and 05/05/2021 - Weapon rebalancing and a somewhat working melee system
I have created a somewhat working melee attack system, although it still needs to be updated. The difficult part was figuring out hit detection. I originally went to a lot of trouble creating a system that involved generating a collider in code (requiring a lot of new skills to be learned) and rotating it to sweep across an attack area, continually checking for targets. This system was very complicated, and I opted to ask in a chat room for advice, at which point I was informed that it was an overly complicated way of doing things. Although disheartened, the mesh creation skills I learned would undoubtedly be valuable in future endeavours.

A common system I was informed of, used by Mordhau, a very popular first-person medieval combat game, was to have a traditional weapon swing animation but add raycasts that shoot a small distance out of the weapon, to detect anything in its swing path. I did not have any animations and didn’t want to create a controller, but I went to the trouble of creating a primitive ‘weapon’ for my player consisting of a cube, and neutral, swing start and swing end positions for it to lerp between, creating a primitive animation. This system is complete to a usable level, and it seems to work, with the player pleasantly swinging their weapon when the button is pressed, and hitting things. Actually hitting enemies proved rather difficult, but this could be a result of range, enemy movement, hitbox size and many other factors I will need to study at some other time. Overall, I am pleased with how this has turned out so far.

I also buffed the player's pistol to improve its critical hit damage and tighten its spread, as it was simply too hard to hit enemies with it (although this might also be because of how easily enemies can dodge attacks).