Update 24/07/2021 – 25/07/2021 – Setting up soldier model animations
Created a class, HumanBodyAnimationData, to handle the rotation of the soldier’s torso and head during runtime. This function works reasonably well, but was a bit complicated to set up, considering I had to include extra euler angle variables to account for the weird angle directions of the transforms for many of the soldier’s rigged bones.

Created a class, AITargetingBeam, to better calculate the beam telegraphing the direction of the enemy’s attack. This involved casting rays to determine where the projectiles would presumably hit, and ensuring the beam faces in that direction specifically and doesn’t clip through walls, instead of shooting directly forward out of the weapon’s barrel (this caused problems since the weapon being held is now animated and moves around in a way that isn’t representative of the aiming direction.

try and fix some problems with the new enemy’s aiming.

I am very pleased with the progress I have made on animating this enemy. However, there are still many things I need to sort out, primarily including:
•	There appears to be some kind of bug where the enemy’s laser sight will not go away, and they will not fire their attack burst.
•	The enemy will not trigger their walking animation until they have reached a certain speed, even if they have already started moving.
•	The enemy walk animation will continue walking forward even when they are backpedalling (also when strafing, but the model I downloaded did not include a strafing animation, so that will be unavoidable).
•	The enemy’s aiming feels awkward and slow to respond, and isn’t very engaging to fight. Previously, I altered the enemy AI targeting system to directly rotate the character’s aim, whereas previously there was a position value that moved towards the character. I created a newer revision of the old aiming system and temporarily implemented it in place of the angular aiming system, to see what changed, but if anything this made the problem worse. I theorise that the best aiming system for my enemies would be some combination of the two systems, but I don’t know how to combine them yet.

On an unrelated note, I also removed the MeleeCollisionDetector class, which was being completely unused and not a part of any system I planned on implementing.