Update 26/07/2021: Polishing up enemy aiming code
I changed the AI’s aiming system again, to the new version. There seem to be major bugs with the aiming code, but I am not able to figure them out yet, as I had other work to do today and didn’t have very much time to work on this. I have a theory as to what the problem is though.
The AI’s head, an empty transform used for the purpose of aiming and viewing, has to rotate in any direction, independently of the direction its NavMeshAgent body moves in. But being a child of the NavMeshAgent, its efforts to continuously look in a particular direction is continually hamstrung by the NavMeshAgent regularly turning around and doing its own thing, altering the head’s direction with it. The old system involved a dedicated Vector3 value named aimMarker, a hypothetical world position that moved through space at a particular speed, and which the AI’s head always snapped to look at. This seemed to not suffer from the problem I’m describing, so I made a Quaternion value to serve a similar purpose. The updated rotation value is stored and applied to the AI’s head. This process, however, did not work. While I’m still not entirely sure what the problem is, it might be something to do with the storing of Quaternion values. I may plan on replacing it with a Vector3 value for direction (as opposed to the old Vector3 value that recorded a position), and store that for the head to look towards. Other factors I need to take into account are if I want my aiming system to stay aiming in a particular direction, or at a particular position, plus when to use world versus local space calculations. I may have to do lots more testing before I reach a result I’m pleased with, but I’m still very proud of my progress so far.

Factors to consider:
Direct rotation versus looking towards a Vector3 that moves linearly through space
*
Aiming in a particular direction versus aiming towards a particular direction
*
World space versus local space
=
Eight potential aiming systems, with varying degrees of successful functionality