Update 29/07/2021 – Solving AI aiming and navigation problems, fixing player crouch controls
I updated the player’s crouching controls slightly. Previously, the player’s collider was in the centre of their body, which made it slightly more annoying to position them in the scene. All of the crouching code also adjusted the appropriate player variables based off this arrangement, so I had to change them accordingly. The new crouching code is somewhat cleaner as well, which is nice.

I am still in the process of fixing my AI’s various issues. I figured out that a big part of the AI’s bugs is due to the movement routines. One pattern I noticed, which I realised was separate from the aiming issues, is that the AI would rapidly shift about from position to position without ever stopping in one place. After monitoring the AI state machine during runtime, and adding code to draw debug lines indicating the NavMeshAgent’s destination, I found out that the enemy was rapidly switching between different destinations, even when it had no need to move. I dug in the code in the behaviour for engaging an acquired target, and I discovered that the AI would accidentally detect its own colliders as a wall, and believe it had lost line of sight, causing it to find a new position. It was unusual, since I believed I had already fixed this issue, but I added code to detect the AI’s various colliders, and ignore them when performing the line of sight check.

This fixed the problem – mostly. While the AI does seem to commit to positions, it will occasionally get stuck in a position where its position rapidly alternates between two points, which both seem to be unreachable. While in this state, its look direction wavers and it takes a very long time for the agent to snap out of it and resume aiming their attacks at the player. I have not yet figured out what is causing this, although it still seems to be coming from the EngageTarget behaviour. An obvious place to start would be to find every point where the agent’s position is assigned, and place a different print message, so I can find out exactly what functions are changing the agent’s destination in such an undesirable way.

I added new variables to the EngageTarget class. I noticed the agent got spooked too easily when the player moved, and adjusted their position too often. I added a second set of variables for a minimum and maximum range, for four total – two when deciding on a position for the agent to move to, and two for deciding if the player has moved to a suboptimal position. This way, I can make the second set of ranges more generous, so the enemy will not move around so much.

I need to update the enemy AI aiming code to specifically prioritise colliders, rather than just aiming towards the player’s transform. Since the player and enemies have their main transforms at the very bottom so they can be easily positioned on a floor, this means enemies will aim at the feet of whoever they are attacking.

I updated the NullableVector3 class used for getting positions on the NavMesh. I got rid of the static New variable, and created proper constructors to replace it. That being said, it may be even better for me to take all functions that produce a NullableVector3, and turn them into bools that alter a pre-established Vector3, like what raycasts do with the RaycastHit variable. But since my current system works well, I have more important things to create.