Update 05/08/2021 – Fixing enemy aiming and line of sight, tweaking damage and gun stats
I tweaked some of the weapon and enemy stats. The pistol was pathetically weak, dealing 3 points of damage when a standard human enemy had 25. I increased its damage and critical hit damage drastically, to kill a standard enemy with 5 headshots or 2 bodyshots. I increased the damage of the assault rifle, to kill a standard sized enemy with a 3-round burst or a single headshot, to make sniping more fun and viable. To keep the weapon from being overpowered, I increased the switching time, to make it harder to deploy in a close combat situation and mainly useful at long ranges. I also noticed the rifleman, the enemy I was using for most of my AI testing, had shockingly low damage, so I more than doubled it so they would pose a legitimate threat.

I fixed the AI aiming so they aim at the center of the player’s colliders, rather than at their feet. The original system would use the centre of the player’s transform as a frame of reference, which isn’t particularly useful since certain objects don’t necessarily have their transforms in the centre of their bodies (usually at their feet so it’s easier to place them on a map). I created a function that gets all the hitboxes of the player or NPC that the AI is targeting, and uses Unity’s Bounds function to check each collider and create a box encompassing all of them. The AI is then able to easily aim at the centre of the box.

I also seem to have solved the jittery AI aiming problems I have described in previous commit logs. I used the debug mode to draw a line connecting the AI’s head to the point they are trying to aim at, in this case, the player. This line should always point at the player if the enemy is in range and has established line of sight. The line proceeded to constantly flicker and die as the enemy’s aim bugged out. After adding a print message to the console, I found out the enemy was continuously losing track of the player despite having no reason to, code brought on by an if sequence returning false. After checking the two parameters in said if statement (if the AI has an assigned target, and if it has line of sight with said target), I found that the line of sight bool uses deprecated old code, that I’d made a replacement version of some time ago. I assumed the old code was causing the AI to repeatedly see its own colliders as a wall and return false, something that again, I thought I’d patched out.

I replaced it with a newer line of sight check I’d made, which constantly returned false. After using a debug ray to check, I confirmed that the line of sight check was aiming for the target’s feet (I realised that in the execution order I was launching the line of sight check before using the fancy new code I mentioned earlier to determine the actual centre of the player), and I realised, to my amusement, that the line of sight check was still being caught… on the thing it was looking for. I updated the list of exception colliders to include the target’s colliders as well as the AI’s, but it still caused the problem. I then realised why.

The new line of sight checking code I was using involves launching a Physics.RaycastAll in the exact direction and distance from the viewer to the target. It then runs a for loop that goes through all colliders caught in the raycast, and with each one performs a secondary for loop that checks them against a list of exception colliders, specified as a parameter when running the function. I realised that the function, when scanning a particular raycast hit, would return the function false because the collider in the raycast hit was different to the specific exception collider it was checked against, even if the collider in the raycast hit was still one of the exceptions.

I then had to rearrange the process. Instead of returning false as soon as it detected a mismatch, each time the primary checking loop brought up a new raycast hit, I declared a bool named hitAnException, then ran the secondary loop. If any of the exceptions matched the collider in the RaycastHit, this bool was set to true and the secondary loop ended prematurely. If this bool was not set to true, an if statement at the end of the secondary loop would trigger, returning the whole function as false, since this would mean an actual obstacle blocked line of sight. The primary loop would then repeat this process with each additional raycast hit until an obstacle was found, or the loop ended returning true.

This was what finally solved the problem, and now my AI seems to be consistently and reliably determining line of sight and shooting at the player. I won’t get excited though, as I thought I’d fixed my AI several times already. I won’t be surprised if it breaks again for some reason. In fact, I think I just noticed a point where the AI had the opposite problem, not having line of sight but refusing to pursue the player. Plus, I’m not entirely sure I’ve fixed the issue of the AI’s aiming being jerked around when the NavMeshAgent makes sharp turns. And I still need to fix my AI’s movement issues.

I also added a dedicated to-do list in the project, to easily keep track of everything I want to work on.