Update 06/08/2021 – Massive fixes to AI aiming code and weapon wheel, streamlined gun firing code, miscellaneous bugfixing
I am making fantastic progress on my AI’s aiming code. I fixed the recurring aiming problem where the AI’s aim would drastically swing around due to its NavMeshAgent making sharp turns, throwing off its accuracy and making it almost impossible for it to hit anything. This is because the aiming code used a ‘head’ transform for directional values, which was unavoidably parented to the body and thus subject to the movements of its parent. This is something that I have finally fixed, and it works very well.

Instead of using the head’s transform rotation as reference for the aiming direction, I changed the code to only use its position. I created a new Quaternion value named lookDirectionQuaternion, which is rotated and referenced during runtime instead of the head. Since it is not attached to a transform, and thus not parented to anything, it will stay in the same direction no matter what direction the NavMeshAgent is moving in. For cosmetic purposes, the head is then rotated to match the Quaternion value in LateUpdate(). I also made several public Vector3 properties representing directions based on lookDirectionQuaternion, to be easily referenced in other scripts without having to perform more calculations on it or reference the head transform.

For resetting the AI’s look position if they are not specifically aiming at anything, I replaced the old ResetLookDirection() function with a new one, named ReturnToNeutralLookPosition(). The old function tries to work like a fire-and-forget function, instantly resetting the look values and the head’s rotation to straight forward, but it is called every frame. One would expect the best course of action to be to update my code so it is only called once when the character stops moving, but I realised I could do better. The problem with trying to reset the look variables in a fire-and-forget function is that the rotation not being tied to the NavMeshAgent’s transform actually works against it, and the AI will continuously look in a single direction. Additionally, instantly resetting the look looks awkward, and would require an IEnumerator to implement smoothly. This new function takes advantage of being called every frame to rotate the AI’s look direction over time to always match the direction its agent is facing. With this, I was also able to finally delete all the old aiming code for being truly redundant.

I also fixed a recurring problem where the enemy AI’s aiming speed seemed too fast, and the player was barely able to dodge their shots after the telegraph. I have code that alters the AI’s aim speed by two different multipliers - once when they first telegraph their attack, and again when they actually start attacking. I found out that the code that was supposed to alter the aim speed was actually altering a different set of variables instead, which I had to fix. I also found out that the code setting the regular aim speed while acquiring a target was being applied every frame, while the code that sets the aim speed while telegraphing and attacking were only being applied once each when appropriate. This meant that the aim speed was being immediately changed back to the way it was before, having no effect. I changed the code to set the default aim speed once when the attack state is first entered, and once when the attack ends – the two times when it was actually necessary.

This fixed the problem, but also created a new one – when the AI telegraphs their attack and doesn’t adjust their aim, they will keep moving and throw their own aim off, almost never hitting the target. I decided to take a different approach, then. I created a public bool named lockOntoTargetWorldPosition. If this is set to true, and the AI has started their attack, they will stop tracking the target and stay focused on the last place it was prior to targeting. After bringing the telegraph and attack aim multipliers up from zero so it would actually continue aiming towards the spot, I seem to have taken care of this problem. That being said, I’ve noticed a problem where despite aiming properly, the AI’s bullets are travelling in strange, inconsistent directions. I did a lot of testing, most notably comparing it against the code for the player, but have not yet found the solution.

Fortunately, looking at my gun firing code allowed me to fix some problems with it. I found out that I’ve been using two separate but virtually identical functions for firing guns, one for the player and one for the enemy. Since the gun firing function used for the enemies was slightly cleaner, I changed the player code to use that as well. I also fixed a bug where the invocation for the UnityEvent that triggers when firing was inside a for loop and activated with every projectile of a shot (resulting in weird effects when the player fired the shotgun).

I also created new code in the AIAttack class, almost identical to the aiming speed adjustment code, that alters the AI’s movement speed while telegraphing and attacking. As well as also assisting in having fair telegraphs for the player, this will allow the AI to move in unique ways while attacking, allowing for even more varied enemy behaviour. This was slightly confusing as I couldn’t think of any clean way to reference my AI’s current state machine behaviours in the main class, but I figured out a solution that involved adding a public field in my AIAttackingState behaviour, referencing the movement behaviour in it and having said attack behaviour give the variable to the attack class when the state starts.

I fixed a bug with the player’s weapon selector, where the icons didn’t appear. Due to some complication with RectTransform.anchoredPosition, the images were instantiated and moved to a point far away from the actual HUD element. I solved this by assigning the object’s local positions and rotations, instead of their world or anchored transform data. I also updated the code used to refresh the weapon wheel with new icons, to reuse any existing icons without destroying them, preventing unnecessary garbage collection.

And finally, I upped the pistol’s critical hit multiplier from 2 to 3. Now it can kill a standard enemy in 2 headshots instead of 3, which I felt was a bit too weak even for a sidearm.